refactor: refactor into `main.js` and `flatter.js`
* remove automatic action input parsing
* move code to `flatter.js`
* only include extra files when uploading GitHub Pages artifact
* add inputs for advanced use cases
* add a license
