+++ action.yml
@@ -1,3 +1,6 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+# SPDX-FileCopyrightText: 2022 Andy Holmes <andrew.g.r.holmes@gmail.com>
+
 name: Flatter
 description: Build and distribute Flatpak applications
 author: Andy Holmes
@@ -44,14 +47,18 @@ inputs:
     description: >
       The fingerprint of a GPG key to sign the repository with.
     required: false
-  install-deps-from:
-    description: >
-      The remote to install build dependencies from.
-    required: false
   repo:
     description: >
       The folder used to store Flatpak repository files.
     default: repo
+    required: false
+  flatpak-builder-args:
+    description: >
+      Extra command-line options for `flatpak-builder`.
+    required: false
+  flatpak-build-bundle-args:
+    description: >
+      Extra command-line options for `flatpak build-bundle`.
     required: false
 runs:
   using: node16
